+++ TessellationAndVoxelizationGeometryLibrary/MarchingCubes/MarchingCubes.VoxelizedSolid.cs
@@ -1,4 +1,3 @@
-using System;
 using TVGL.Voxelization;
 
 namespace TVGL
@@ -6,6 +5,7 @@ namespace TVGL
     internal class MarchingCubesDenseVoxels : MarchingCubes<VoxelizedSolid, bool>
     {
         private readonly double coordToVoxelIndex;
+
         internal MarchingCubesDenseVoxels(VoxelizedSolid solid, int numVoxelsPerGrid)
             : base(solid, solid.VoxelSideLength * numVoxelsPerGrid)
 
@@ -37,19 +37,21 @@ namespace TVGL
             switch (direction)
             {
                 case 0:
-                    var maxX = (int)(to.X - from.X);
+                    var maxX = to.X - from.X;
                     for (int i = 1; i < maxX; i++)
                         if (GetValueFromSolid((from.X + sign * i), from.Y, from.Z) != GetValueFromSolid((from.X + sign * (i - 1)), from.Y, from.Z))
                             return gridToCoordinateFactor * (i + 0.5) / maxX;
                     break;
+
                 case 1:
-                    var maxY = (int)(to.Y - from.Y);
+                    var maxY = to.Y - from.Y;
                     for (int i = 1; i < maxY; i++)
                         if (GetValueFromSolid(from.X, (from.Y + sign * i), from.Z) != GetValueFromSolid(from.X, (from.Y + sign * (i - 1)), from.Z))
                             return gridToCoordinateFactor * (i + 0.5) / maxY;
                     break;
+
                 case 2:
-                    var maxZ = (int)(to.X - from.X);
+                    var maxZ = to.X - from.X;
                     for (int i = 1; i < maxZ; i++)
                         if (GetValueFromSolid(from.X, from.Y, (from.Z + sign * i)) != GetValueFromSolid(from.X, from.Y, (from.Z + sign * (i - 1))))
                             return gridToCoordinateFactor * (i + 0.5) / maxZ;
